fixed icons bugs

+++ Utils/GetIconUseDll.cs
@@ -1,10 +1,9 @@
 ﻿using System;
-using System.CodeDom.Compiler;
-using System.Collections.Generic;
 using System.Drawing;
 using System.IO;
-using System.Linq;
 using System.Runtime.InteropServices;
+
+// folder icon code taken from here: https://stackoverflow.com/questions/42910628/is-there-a-way-to-get-the-windows-default-folder-icon-using-c
 
 // based on code from here: https://www.codeproject.com/Articles/390514/Playing-with-a-MVVM-Tabbed-TreeView-for-a-File-Exp
 
@@ -23,8 +22,33 @@ namespace mouse_tracking_web_app.Utils
 {
     public class ShellIcon
     {
+        private class Win32
+        {
+            public const uint SHGFI_ICON = 0x100;
+            public const uint SHGFI_LARGEICON = 0x0; // Large icon
+            public const uint SHGFI_SMALLICON = 0x1; // Small icon
+            public const uint SHSIID_FOLDER = 0x3;  // Folder icon
+            public const uint USEFILEATTRIBUTES = 0x000000010; // when the full path isn't available
+
+            [DllImport("User32.dll")]
+            public static extern int DestroyIcon(IntPtr hIcon);
+
+            [DllImport("shell32.dll")]
+            public static extern int SHGetStockIconInfo(uint siid, uint uFlags, ref SHSTOCKICONINFO psii);
+
+            //extra
+            [DllImport("Shell32.dll")]
+            public static extern int ExtractIconEx(string libName, int iconIndex, IntPtr[] largeIcon, IntPtr[] smallIcon, uint nIcons);
+
+            [DllImport("shell32.dll")]
+            public static extern IntPtr SHGetFileInfo(string pszPath, uint dwFileAttributes, ref SHFILEINFO psfi, uint cbSizeFileInfo, uint uFlags);
+        }
+
+
         public ShellIcon()
         { }
+
+        public static Icon FolderLarge => GetStockIcon(Win32.SHSIID_FOLDER, Win32.SHGFI_LARGEICON);
 
         public static Icon GetLargeIcon(string fileName)
         {
@@ -36,16 +60,23 @@ namespace mouse_tracking_web_app.Utils
                 _ = Win32.DestroyIcon(shinfo.hIcon);
                 return icon;
             }
+            // can't create a icon
             catch
             {
-                //    string tempDirectory = @"c:\\temp";
-                //    TempFileCollection coll = new TempFileCollection(tempDirectory, true);
-                    string[] sName = fileName.Split('.');
-                //    string tempFilename = coll.AddExtension(sName[sName.Length - 1], true);
-                string tempFileName = Path.Combine(".", $"temp.{sName[sName.Length - 1]}");
+                // if the path is a directory, simply return directory icon
+                if (File.GetAttributes(fileName).HasFlag(FileAttributes.Directory))
+                    //return GetStockIcon(SHSIID_FOLDER, SHGSI_LARGEICON);
+                    return FolderLarge;
+
+                // if not a directory, copy the file somewhere else and use the extention
+                // TODO: maybe I should save some kind of cache of all file types
+                string[] sName = fileName.Split('.');
+                string tempFileName;
+                tempFileName = Path.Combine(".", $"temp.{sName[sName.Length - 1]}");
                 File.WriteAllText(tempFileName, "Hello World");
 
                 _ = Win32.SHGetFileInfo(tempFileName, 0, ref shinfo, (uint)Marshal.SizeOf(shinfo), Win32.SHGFI_ICON | Win32.SHGFI_LARGEICON);
+                
                 try
                 {
                     Icon icon = (Icon)Icon.FromHandle(shinfo.hIcon).Clone();
@@ -59,7 +90,8 @@ namespace mouse_tracking_web_app.Utils
                 }
                 finally
                 {
-                    File.Delete(tempFileName);
+                    if (!string.IsNullOrEmpty(tempFileName))
+                        File.Delete(tempFileName);
                 }
             }
         }
@@ -71,6 +103,19 @@ namespace mouse_tracking_web_app.Utils
             //The icon is returned in the hIcon member of the shinfo struct
             Icon icon = (Icon)Icon.FromHandle(shinfo.hIcon).Clone();
             _ = Win32.DestroyIcon(shinfo.hIcon);
+            return icon;
+        }
+
+        private static Icon GetStockIcon(uint type, uint size)
+        {
+            var info = new SHSTOCKICONINFO();
+            info.cbSize = (uint)Marshal.SizeOf(info);
+
+            Win32.SHGetStockIconInfo(type, Win32.SHGFI_ICON | size, ref info);
+
+            var icon = (Icon)Icon.FromHandle(info.hIcon).Clone(); // Get a copy that doesn't use the original handle
+            Win32.DestroyIcon(info.hIcon); // Clean up native icon to prevent resource leak
+
             return icon;
         }
 
@@ -88,22 +133,16 @@ namespace mouse_tracking_web_app.Utils
             public string szTypeName;
         };
 
-        private class Win32
+        [StructLayout(LayoutKind.Sequential, CharSet = CharSet.Unicode)]
+        public struct SHSTOCKICONINFO
         {
-            public const uint SHGFI_ICON = 0x100;
-            public const uint SHGFI_LARGEICON = 0x0; // Large icon
-            public const uint SHGFI_SMALLICON = 0x1; // Small icon
-            public const uint USEFILEATTRIBUTES = 0x000000010; // when the full path isn't available
+            public uint cbSize;
+            public IntPtr hIcon;
+            public int iSysIconIndex;
+            public int iIcon;
 
-            [DllImport("User32.dll")]
-            public static extern int DestroyIcon(IntPtr hIcon);
-
-            //extra
-            [DllImport("Shell32.dll")]
-            public static extern int ExtractIconEx(string libName, int iconIndex, IntPtr[] largeIcon, IntPtr[] smallIcon, uint nIcons);
-
-            [DllImport("shell32.dll")]
-            public static extern IntPtr SHGetFileInfo(string pszPath, uint dwFileAttributes, ref SHFILEINFO psfi, uint cbSizeFileInfo, uint uFlags);
+            [MarshalAs(UnmanagedType.ByValTStr, SizeConst = 260)]
+            public string szPath;
         }
     }
 }
